Refactor dependency injection for easy manage ablility of configure services
