+++ src/Depreciated/IEngineFactory.cs
@@ -1,10 +1,17 @@
 // Copyright (c) 2014-2020 Sarin Na Wangkanai, All Rights Reserved.
 // The Apache v2. See License.txt in the project root for license information.
 
+using System;
+
+using Wangkanai.Detection.Models;
+
 namespace Wangkanai.Detection
 {
-    public interface IEngineResolver : IResolver
+    [Obsolete]
+    public interface IEngineFactory
     {
-        IEngineFactory Engine { get; }
+        string Maker { get; }
+        Version Version { get; }
+        Engine Type { get; }
     }
 }
