+++ src/Depreciated/IPlatformResolver.cs
@@ -1,8 +1,11 @@
 // Copyright (c) 2014-2020 Sarin Na Wangkanai, All Rights Reserved.
 // The Apache v2. See License.txt in the project root for license information.
 
+using System;
+
 namespace Wangkanai.Detection
 {
+    [Obsolete]
     public interface IPlatformResolver : IResolver
     {
         IPlatformFactory Platform { get; }
